Move all map metadata properties onto a Mixin class
