+++ sunpy/map/mapbase.py
@@ -1,7 +1,6 @@
 """
 Map is a generic Map class from which all other Map classes inherit from.
 """
-import re
 import copy
 import html
 import inspect
@@ -10,7 +9,6 @@ import textwrap
 import itertools
 import webbrowser
 from tempfile import NamedTemporaryFile
-from collections import namedtuple
 
 import matplotlib
 import matplotlib.pyplot as plt
@@ -32,36 +30,29 @@ from astropy.utils.metadata import MetaData
 from astropy.visualization import AsymmetricPercentileInterval, HistEqStretch, ImageNormalize
 from astropy.visualization.wcsaxes import Quadrangle, WCSAxes
 
+import sunpy
 # The next two are not used but are called to register functions with external modules
-import sunpy.coordinates
 import sunpy.io as io
-import sunpy.io._fits
-import sunpy.visualization.colormaps
-from sunpy import config, log
-from sunpy.coordinates import HeliographicCarrington, get_earth, sun
+from sunpy import config
 from sunpy.coordinates.utils import get_rectangle_coordinates
 from sunpy.image.resample import resample as sunpy_image_resample
 from sunpy.image.resample import reshape_image_to_4d_superpixel
 from sunpy.image.transform import _get_transform_method, _rotation_function_names, affine_transform
-from sunpy.sun import constants
-from sunpy.time import is_time, parse_time
-from sunpy.util import MetaDict, expand_list
+from sunpy.util import MetaDict
 from sunpy.util.decorators import (
     add_common_docstring,
     cached_property_based_on,
     check_arithmetic_compatibility,
 )
-from sunpy.util.exceptions import warn_metadata, warn_user
+from sunpy.util.exceptions import warn_user
 from sunpy.util.functools import seconddispatch
 from sunpy.util.util import _figure_to_base64, fix_duplicate_notes
 from sunpy.visualization import axis_labels_from_ctype, peek_show, wcsaxes_compat
 from sunpy.visualization.colormaps import cm as sunpy_cm
+from .mixins.mapmeta import MapMetaMixin, MapMetaValidationError, PixelPair
 
 TIME_FORMAT = config.get("general", "time_format")
-PixelPair = namedtuple('PixelPair', 'x y')
-SpatialPair = namedtuple('SpatialPair', 'axis1 axis2')
 _NUMPY_COPY_IF_NEEDED = False if np.__version__.startswith("1.") else None
-_META_FIX_URL = 'https://docs.sunpy.org/en/stable/how_to/fix_map_metadata.html'
 
 # Manually specify the ``.meta`` docstring. This is assigned to the .meta
 # class attribute in GenericMap.__init__()
@@ -106,11 +97,7 @@ to the standard PC_ij described in section 6.1 of .
 __all__ = ['GenericMap', 'MapMetaValidationError', 'PixelPair']
 
 
-class MapMetaValidationError(AttributeError):
-    pass
-
-
-class GenericMap(NDData):
+class GenericMap(MapMetaMixin, NDData):
     """
     A Generic spatially-aware 2D data array
 
@@ -579,10 +566,6 @@ class GenericMap(NDData):
         new_data = value / self.quantity
         return self._new_instance_from_op(new_data)
 
-    @property
-    def _meta_hash(self):
-        return self.meta.item_hash()
-
     def _set_symmetric_vmin_vmax(self):
         """
         Set symmetric vmin and vmax about zero
@@ -590,72 +573,6 @@ class GenericMap(NDData):
         threshold = np.nanmax(abs(self.data))
         self.plot_settings['norm'].vmin = -threshold
         self.plot_settings['norm'].vmax = threshold
-
-    @property
-    @cached_property_based_on('_meta_hash')
-    def wcs(self):
-        """
-        The `~astropy.wcs.WCS` property of the map.
-        """
-        w2 = astropy.wcs.WCS(naxis=2)
-
-        # Add one to go from zero-based to one-based indexing
-        w2.wcs.crpix = u.Quantity(self.reference_pixel) + 1 * u.pix
-        # Make these a quantity array to prevent the numpy setting element of
-        # array with sequence error.
-        # Explicitly call ``.to()`` to check that scale is in the correct units
-        w2.wcs.cdelt = u.Quantity([self.scale[0].to(self.spatial_units[0] / u.pix),
-                                   self.scale[1].to(self.spatial_units[1] / u.pix)])
-        w2.wcs.crval = u.Quantity([self._reference_longitude, self._reference_latitude])
-        w2.wcs.ctype = self.coordinate_system
-        w2.wcs.pc = self.rotation_matrix
-        w2.wcs.set_pv(self._pv_values)
-        # FITS standard doesn't allow both PC_ij *and* CROTA keywords
-        w2.wcs.crota = (0, 0)
-        w2.wcs.cunit = self.spatial_units
-        w2.wcs.dateobs = self.date.isot
-        w2.wcs.aux.rsun_ref = self.rsun_meters.to_value(u.m)
-
-        # Set observer coordinate information except when we know it is not appropriate (e.g., HGS)
-        sunpy_frame = sunpy.coordinates.wcs_utils._sunpy_frame_class_from_ctypes(w2.wcs.ctype)
-        if sunpy_frame is None or hasattr(sunpy_frame, 'observer'):
-            # Clear all the aux information that was set earlier. This is to avoid
-            # issues with maps that store multiple observer coordinate keywords.
-            # Note that we have to create a new WCS as it's not possible to modify
-            # wcs.wcs.aux in place.
-            header = w2.to_header()
-            for kw in ['crln_obs', 'dsun_obs', 'hgln_obs', 'hglt_obs']:
-                header.pop(kw, None)
-            w2 = astropy.wcs.WCS(header)
-
-            # Get observer coord, and set the aux information
-            obs_coord = self.observer_coordinate
-            sunpy.coordinates.wcs_utils._set_wcs_aux_obs_coord(w2, obs_coord)
-
-        # Set the shape of the data array
-        w2.array_shape = self.data.shape
-
-        # Validate the WCS here.
-        w2.wcs.set()
-        return w2
-
-    @property
-    def coordinate_frame(self):
-        """
-        An `astropy.coordinates.BaseCoordinateFrame` instance created from the coordinate
-        information for this Map, or None if the frame cannot be determined.
-        """
-        try:
-            return astropy.wcs.utils.wcs_to_celestial_frame(self.wcs)
-        except ValueError as e:
-            warn_user(f'Could not determine coordinate frame from map metadata.\n{e}')
-            return None
-
-    @property
-    def _coordinate_frame_name(self):
-        if self.coordinate_frame is None:
-            return 'Unknown'
-        return self.coordinate_frame.name
 
     def _as_mpl_axes(self):
         """
@@ -724,297 +641,6 @@ class GenericMap(NDData):
         """
         return np.nanmax(self.data, *args, **kwargs)
 
-    @staticmethod
-    def _parse_fits_unit(unit_str):
-        replacements = {'gauss': 'G',
-                        'dn': 'ct',
-                        'dn/s': 'ct/s',
-                        'counts / pixel': 'ct/pix',}
-        if unit_str.lower() in replacements:
-            unit_str = replacements[unit_str.lower()]
-        unit = u.Unit(unit_str, format='fits', parse_strict='silent')
-        if isinstance(unit, u.UnrecognizedUnit):
-            warn_metadata(f'Could not parse unit string "{unit_str}" as a valid FITS unit.\n'
-                          f'See {_META_FIX_URL} for how to fix metadata before loading it '
-                          'with sunpy.map.Map.\n'
-                          'See https://fits.gsfc.nasa.gov/fits_standard.html for '
-                          'the FITS unit standards.')
-            unit = None
-        return unit
-
-    @property
-    def unit(self):
-        """
-        Unit of the map data.
-
-        This is taken from the 'BUNIT' FITS keyword. If no 'BUNIT' entry is
-        present in the metadata then this returns `None`. If the 'BUNIT' value
-        cannot be parsed into a unit a warning is raised, and `None` returned.
-        """
-        unit_str = self.meta.get('bunit', None)
-        if unit_str is None:
-            return
-
-        return self._parse_fits_unit(unit_str)
-
-# #### Keyword attribute and other attribute definitions #### #
-
-    def _base_name(self):
-        """Abstract the shared bit between name and latex_name"""
-        if self.measurement is None:
-            format_str = "{nickname} {date}"
-        else:
-            format_str = "{nickname} {{measurement}} {date}"
-        return format_str.format(nickname=self.nickname,
-                                 date=parse_time(self.date).strftime(TIME_FORMAT))
-
-    @property
-    def name(self):
-        """Human-readable description of the Map."""
-        return self._base_name().format(measurement=self.measurement)
-
-    @property
-    def latex_name(self):
-        """LaTeX formatted description of the Map."""
-        if isinstance(self.measurement, u.Quantity):
-            return self._base_name().format(measurement=self.measurement._repr_latex_())
-        else:
-            return self.name
-
-    @property
-    def nickname(self):
-        """An abbreviated human-readable description of the map-type; part of
-        the Helioviewer data model."""
-        return self._nickname if self._nickname else self.detector
-
-    @nickname.setter
-    def nickname(self, n):
-        self._nickname = n
-
-    def _get_date(self, key):
-        time = self.meta.get(key, None)
-        if not time:
-            return
-
-        # Get the time scale
-        if 'TAI' in time:
-            # SDO specifies the 'TAI' scale in their time string, which is parsed
-            # by parse_time(). If a different timescale is also present, warn the
-            # user that it will be ignored.
-            timesys = 'TAI'
-            timesys_meta = self.meta.get('timesys', '').upper()
-            if timesys_meta not in ('', 'TAI'):
-                warn_metadata('Found "TAI" in time string, ignoring TIMESYS keyword '
-                              f'which is set to "{timesys_meta}".')
-        else:
-            timesys = self._timesys
-
-        return parse_time(time, scale=timesys.lower())
-
-    @property
-    def _timesys(self):
-        """
-        Time system.
-        """
-        # UTC is the FITS standard default
-        return self.meta.get('timesys', 'UTC')
-
-    @property
-    def date_start(self):
-        """
-        Time of the beginning of the image acquisition.
-
-        Taken from the DATE-BEG FITS keyword.
-        """
-        return self._get_date('date-beg')
-
-    @property
-    def date_end(self):
-        """
-        Time of the end of the image acquisition.
-
-        Taken from the DATE-END FITS keyword.
-        """
-        return self._get_date('date-end')
-
-    @property
-    def date_average(self):
-        """
-        Average time of the image acquisition.
-
-        Taken from the DATE-AVG FITS keyword if present, otherwise halfway
-        between `date_start` and `date_end` if both pieces of metadata are
-        present.
-        """
-        avg = self._get_date('date-avg')
-        if avg is None:
-            start, end = self.date_start, self.date_end
-            if start is not None and end is not None:
-                avg = start + (end - start) / 2
-
-        return avg
-
-    @property
-    def _date_obs(self):
-        # Get observation date from date-obs, falling back to date_obs
-        time = self._get_date('date-obs')
-        if is_time(self.meta.get('date_obs', None)):
-            time = time or self._get_date('date_obs')
-        return time
-
-    @property
-    def date(self):
-        """
-        Image observation time.
-
-        For different combinations of map metadata this can return either
-        the start time, end time, or a time between these. It is recommended
-        to use `~sunpy.map.GenericMap.date_average`,
-        `~sunpy.map.GenericMap.date_start`, or `~sunpy.map.GenericMap.date_end`
-        instead if you need one of these specific times.
-
-        Taken from, in order of preference:
-
-        1. The DATE-OBS FITS keyword
-        2. `~sunpy.map.GenericMap.date_average`
-        3. `~sunpy.map.GenericMap.date_start`
-        4. `~sunpy.map.GenericMap.date_end`
-        5. The current time
-        """
-        time = self._date_obs
-        time = time or self.date_average
-        time = time or self.date_start
-        time = time or self.date_end
-
-        if time is None:
-            if self._default_time is None:
-                warn_metadata("Missing metadata for observation time, "
-                              "setting observation time to current time. "
-                              "Set the 'DATE-AVG' FITS keyword to prevent this warning.")
-                self._default_time = parse_time('now')
-            time = self._default_time
-
-        return time
-
-    @property
-    def detector(self):
-        """
-        Detector name.
-
-        This is taken from the 'DETECTOR' FITS keyword.
-        """
-        return self.meta.get('detector', "")
-
-    @property
-    def timeunit(self):
-        """
-        The `~astropy.units.Unit` of the exposure time of this observation.
-
-        Taken from the "TIMEUNIT" FITS keyword, and defaults to seconds (as per)
-        the FITS standard).
-        """
-        return u.Unit(self.meta.get('timeunit', 's'))
-
-    @property
-    def exposure_time(self):
-        """
-        Exposure time of the image.
-
-        This is taken from the 'XPOSURE' keyword or the 'EXPTIME' FITS keyword,
-        in that order.
-        """
-        exptime = self.meta.get('xposure') or self.meta.get('exptime')
-        if exptime is not None:
-            return exptime * self.timeunit
-
-    @property
-    def instrument(self):
-        """Instrument name."""
-        return self.meta.get('instrume', "").replace("_", " ")
-
-    @property
-    def measurement(self):
-        """
-        Measurement wavelength.
-
-        This is taken from the 'WAVELNTH' FITS keywords. If the keyword is not
-        present, defaults to `None`. If 'WAVEUNIT' keyword isn't present,
-        defaults to dimensionless units.
-        """
-        return self.wavelength
-
-    @property
-    def waveunit(self):
-        """
-        The `~astropy.units.Unit` of the wavelength of this observation.
-
-        This is taken from the 'WAVEUNIT' FITS keyword. If the keyword is not
-        present, defaults to `None`
-        """
-        if 'waveunit' in self.meta:
-            return u.Unit(self.meta['waveunit'])
-        else:
-            wunit = sunpy.io._fits.extract_waveunit(self.meta)
-            if wunit is not None:
-                return u.Unit(wunit)
-
-    @property
-    def wavelength(self):
-        """
-        Wavelength of the observation.
-
-        This is taken from the 'WAVELNTH' FITS keywords. If the keyword is not
-        present, defaults to `None`. If 'WAVEUNIT' keyword isn't present,
-        defaults to dimensionless units.
-        """
-        if 'wavelnth' in self.meta:
-            return u.Quantity(self.meta['wavelnth'], self.waveunit)
-
-    @property
-    def observatory(self):
-        """
-        Observatory or Telescope name.
-
-        This is taken from the 'OBSRVTRY' FITS keyword.
-        """
-        return self.meta.get('obsrvtry',
-                             self.meta.get('telescop', "")).replace("_", " ")
-
-    @property
-    def processing_level(self):
-        """
-        Returns the FITS processing level if present.
-
-        This is taken from the 'LVL_NUM' FITS keyword.
-        """
-        return self.meta.get('lvl_num', None)
-
-    @property
-    def bottom_left_coord(self):
-        """
-        The physical coordinate at the center of the bottom left ([0, 0]) pixel.
-        """
-        return self.wcs.pixel_to_world(0, 0)
-
-    @property
-    def top_right_coord(self):
-        """
-        The physical coordinate at the center of the the top right ([-1, -1]) pixel.
-        """
-        top_right = np.array([self.dimensions.x.value, self.dimensions.y.value]) - 1
-        return self.wcs.pixel_to_world(*top_right)
-
-    @property
-    def center(self):
-        """
-        Return a coordinate object for the center pixel of the array.
-
-        If the array has an even number of pixels in a given dimension,
-        the coordinate returned lies on the edge between the two central pixels.
-        """
-        center = (np.array([self.dimensions.x.value, self.dimensions.y.value]) - 1) / 2.
-        return self.wcs.pixel_to_world(*center)
-
     @u.quantity_input
     def shift_reference_coord(self, axis1: u.deg, axis2: u.deg):
         """
@@ -1048,355 +674,53 @@ class GenericMap(NDData):
 
         return new_map
 
-    def _rsun_meters(self, dsun=None):
-        """
-        This property exists to avoid circular logic in constructing the
-        observer coordinate, by allowing a custom 'dsun' to be specified,
-        instead of one extracted from the `.observer_coordinate` property.
-        """
-        rsun = self.meta.get('rsun_ref', None)
-        if rsun is not None:
-            return rsun * u.m
-        elif self._rsun_obs_no_default is not None:
-            if dsun is None:
-                dsun = self.dsun
-            return sun._radius_from_angular_radius(self.rsun_obs, dsun)
-        else:
-            log.info("Missing metadata for solar radius: assuming "
-                     "the standard radius of the photosphere.")
-            return constants.radius
-
-    @property
-    def rsun_meters(self):
-        """
-        Assumed radius of observed emission from the Sun center.
-
-        This is taken from the RSUN_REF FITS keyword, if present.
-        If not, and angular radius metadata is present, it is calculated from
-        `~sunpy.map.GenericMap.rsun_obs` and `~sunpy.map.GenericMap.dsun`.
-        If neither pieces of metadata are present, defaults to the standard
-        photospheric radius.
-        """
-        return self._rsun_meters()
-
-    @property
-    def _rsun_obs_no_default(self):
-        """
-        Get the angular radius value from FITS keywords without defaulting.
-        Exists to avoid circular logic in `rsun_meters()` above.
-        """
-        return self.meta.get('rsun_obs',
-                             self.meta.get('solar_r',
-                                           self.meta.get('radius',
-                                                         None)))
-
-    @property
-    def rsun_obs(self):
-        """
-        Angular radius of the observation from Sun center.
-
-        This value is taken (in order of preference) from the 'RSUN_OBS',
-        'SOLAR_R', or 'RADIUS' FITS keywords. If none of these keys are present,
-        the angular radius is calculated from
-        `~sunpy.map.GenericMap.rsun_meters` and `~sunpy.map.GenericMap.dsun`.
-        """
-        rsun_arcseconds = self._rsun_obs_no_default
-
-        if rsun_arcseconds is not None:
-            return rsun_arcseconds * u.arcsec
-        else:
-            return sun._angular_radius(self.rsun_meters, self.dsun)
-
-    @property
-    def coordinate_system(self):
-        """
-        Coordinate system used for x and y axes (ctype1/2).
-
-        If not present, defaults to (HPLN-TAN, HPLT-TAN), and emits a warning.
-        """
-        ctype1 = self.meta.get('ctype1', None)
-        if not ctype1:
-            warn_metadata("Missing CTYPE1 from metadata, assuming CTYPE1 is HPLN-TAN")
-            ctype1 = 'HPLN-TAN'
-
-        ctype2 = self.meta.get('ctype2', None)
-        if not ctype2:
-            warn_metadata("Missing CTYPE2 from metadata, assuming CTYPE2 is HPLT-TAN")
-            ctype2 = 'HPLT-TAN'
-
-        # Astropy WCS does not understand the SOHO default of "solar-x" and
-        # "solar-y" ctypes.  This overrides the default assignment and
-        # changes it to a ctype that is understood.  See Thompson, 2006, A.&A.,
-        # 449, 791.
-        if ctype1.lower() in ("solar-x", "solar_x"):
-            ctype1 = 'HPLN-TAN'
-
-        if ctype2.lower() in ("solar-y", "solar_y"):
-            ctype2 = 'HPLT-TAN'
-
-        return SpatialPair(ctype1, ctype2)
-
-    @property
-    def _supported_observer_coordinates(self):
-        """
-        A list of supported coordinate systems.
-
-        This is a list so it can easily maintain a strict order. The list of
-        two element tuples, the first item in the tuple is the keys that need
-        to be in the header to use this coordinate system and the second is the
-        kwargs to SkyCoord.
-        """
-        return [(('hgln_obs', 'hglt_obs', 'dsun_obs'), {'lon': self.meta.get('hgln_obs'),
-                                                        'lat': self.meta.get('hglt_obs'),
-                                                        'radius': self.meta.get('dsun_obs'),
-                                                        'unit': (u.deg, u.deg, u.m),
-                                                        'frame': "heliographic_stonyhurst"}),
-                (('crln_obs', 'crlt_obs', 'dsun_obs'), {'lon': self.meta.get('crln_obs'),
-                                                        'lat': self.meta.get('crlt_obs'),
-                                                        'radius': self.meta.get('dsun_obs'),
-                                                        'unit': (u.deg, u.deg, u.m),
-                                                        'frame': "heliographic_carrington"}), ]
-
-    @property
-    def _default_observer_coordinate(self):
-        """
-        The default observer coordinate to use when there is insufficient information
-        in the metadata. This should be overridden by map sources as appropriate.
-        """
-
-    def _remove_existing_observer_location(self):
-        """
-        Remove all keys that this map might use for observer location.
-        """
-        all_keys = expand_list([e[0] for e in self._supported_observer_coordinates])
-        for key in all_keys:
-            self.meta.pop(key)
-
     @property
     @cached_property_based_on('_meta_hash')
-    def observer_coordinate(self):
+    def wcs(self):
         """
-        The Heliographic Stonyhurst Coordinate of the observer.
+        The `~astropy.wcs.WCS` property of the map.
         """
-        warning_message = []
-        for keys, kwargs in self._supported_observer_coordinates:
-            missing_keys = set(keys) - self.meta.keys()
-            if not missing_keys:
-                sc = SkyCoord(obstime=self.date, **kwargs)
-                # If the observer location is supplied in Carrington coordinates,
-                # the coordinate's `observer` attribute should be set to "self"
-                if isinstance(sc.frame, HeliographicCarrington):
-                    sc.frame._observer = "self"
+        w2 = astropy.wcs.WCS(naxis=2)
 
-                sc = sc.heliographic_stonyhurst
-                # We set rsun after constructing the coordinate, as we need
-                # the observer-Sun distance (sc.radius) to calculate this, which
-                # may not be provided directly in metadata (if e.g. the
-                # observer coordinate is specified in a cartesian
-                # representation)
-                return SkyCoord(sc.replicate(rsun=self._rsun_meters(sc.radius)))
-            elif missing_keys != keys:
-                frame = kwargs['frame'] if isinstance(kwargs['frame'], str) else kwargs['frame'].name
-                warning_message.append(f"For frame '{frame}' the following metadata is missing: "
-                                       f"{','.join(missing_keys)}")
+        # Add one to go from zero-based to one-based indexing
+        w2.wcs.crpix = u.Quantity(self.reference_pixel) + 1 * u.pix
+        # Make these a quantity array to prevent the numpy setting element of
+        # array with sequence error.
+        # Explicitly call ``.to()`` to check that scale is in the correct units
+        w2.wcs.cdelt = u.Quantity([self.scale[0].to(self.spatial_units[0] / u.pix),
+                                   self.scale[1].to(self.spatial_units[1] / u.pix)])
+        w2.wcs.crval = u.Quantity([self._reference_longitude, self._reference_latitude])
+        w2.wcs.ctype = self.coordinate_system
+        w2.wcs.pc = self.rotation_matrix
+        w2.wcs.set_pv(self._pv_values)
+        # FITS standard doesn't allow both PC_ij *and* CROTA keywords
+        w2.wcs.crota = (0, 0)
+        w2.wcs.cunit = self.spatial_units
+        w2.wcs.dateobs = self.date.isot
+        w2.wcs.aux.rsun_ref = self.rsun_meters.to_value(u.m)
 
-        default = self._default_observer_coordinate
-        if default is not None:
-            # If a map source specifies a default observer, we log a message at the debug level
-            warning_message = (["Missing metadata for observer: assuming custom default observer."]
-                               + warning_message)
-            log.debug("\n".join(warning_message))
-            return default
-        else:
-            # If a map source does not specify a default observer, we assume Earth center and warn
-            warning_message = (["Missing metadata for observer: assuming Earth-based observer."]
-                               + warning_message + [""])
-            warn_metadata("\n".join(warning_message), stacklevel=3)
-            return get_earth(self.date)
+        # Set observer coordinate information except when we know it is not appropriate (e.g., HGS)
+        sunpy_frame = sunpy.coordinates.wcs_utils._sunpy_frame_class_from_ctypes(w2.wcs.ctype)
+        if sunpy_frame is None or hasattr(sunpy_frame, 'observer'):
+            # Clear all the aux information that was set earlier. This is to avoid
+            # issues with maps that store multiple observer coordinate keywords.
+            # Note that we have to create a new WCS as it's not possible to modify
+            # wcs.wcs.aux in place.
+            header = w2.to_header()
+            for kw in ['crln_obs', 'dsun_obs', 'hgln_obs', 'hglt_obs']:
+                header.pop(kw, None)
+            w2 = astropy.wcs.WCS(header)
 
-    @property
-    def heliographic_latitude(self):
-        """Observer heliographic latitude."""
-        return self.observer_coordinate.lat
+            # Get observer coord, and set the aux information
+            obs_coord = self.observer_coordinate
+            sunpy.coordinates.wcs_utils._set_wcs_aux_obs_coord(w2, obs_coord)
 
-    @property
-    def heliographic_longitude(self):
-        """Observer heliographic longitude."""
-        return self.observer_coordinate.lon
+        # Set the shape of the data array
+        w2.array_shape = self.data.shape
 
-    @property
-    def carrington_latitude(self):
-        """Observer Carrington latitude."""
-        hgc_frame = HeliographicCarrington(observer=self.observer_coordinate, obstime=self.date,
-                                           rsun=self.rsun_meters)
-        return self.observer_coordinate.transform_to(hgc_frame).lat
-
-    @property
-    def carrington_longitude(self):
-        """Observer Carrington longitude."""
-        hgc_frame = HeliographicCarrington(observer=self.observer_coordinate, obstime=self.date,
-                                           rsun=self.rsun_meters)
-        return self.observer_coordinate.transform_to(hgc_frame).lon
-
-    @property
-    def dsun(self):
-        """Observer distance from the center of the Sun."""
-        return self.observer_coordinate.radius.to('m')
-
-    @property
-    def _reference_longitude(self):
-        """
-        FITS-WCS compatible longitude. Used in self.wcs and
-        self.reference_coordinate.
-        """
-        return self.meta.get('crval1', 0.) * self.spatial_units[0]
-
-    @property
-    def _reference_latitude(self):
-        return self.meta.get('crval2', 0.) * self.spatial_units[1]
-
-    @property
-    def reference_coordinate(self):
-        """Reference point WCS axes in data units (i.e. crval1, crval2). This value
-        includes a shift if one is set."""
-        return SkyCoord(self._reference_longitude,
-                        self._reference_latitude,
-                        frame=self.coordinate_frame)
-
-    @property
-    def reference_pixel(self):
-        """
-        Pixel of reference coordinate.
-
-        The pixel returned uses zero-based indexing, so will be 1 pixel less
-        than the FITS CRPIX values.
-        """
-        naxis1 = self.meta.get('naxis1', self.data.shape[1])
-        naxis2 = self.meta.get('naxis2', self.data.shape[0])
-        return PixelPair((self.meta.get('crpix1', (naxis1 + 1) / 2.) - 1) * u.pixel,
-                         (self.meta.get('crpix2', (naxis2 + 1) / 2.) - 1) * u.pixel)
-
-    @property
-    def scale(self):
-        """
-        Image scale along the x and y axes in units/pixel
-        (i.e. cdelt1, cdelt2).
-
-        If the CDij matrix is defined but no CDELTi values are explicitly defined,
-        effective CDELTi values are constructed from the CDij matrix.  The effective
-        CDELTi values are chosen so that each row of the PCij matrix has unity norm.
-        This choice is optimal if the PCij matrix is a pure rotation matrix, but may not
-        be as optimal if the PCij matrix includes any skew.
-        """
-        if 'cd1_1' in self.meta and 'cdelt1' not in self.meta and 'cdelt2' not in self.meta:
-            cdelt1 = np.sqrt(self.meta['cd1_1']**2 + self.meta['cd1_2']**2)
-            cdelt2 = np.sqrt(self.meta['cd2_1']**2 + self.meta['cd2_2']**2)
-        else:
-            cdelt1 = self.meta.get('cdelt1', 1.)
-            cdelt2 = self.meta.get('cdelt2', 1.)
-
-        return SpatialPair(cdelt1 * self.spatial_units[0] / u.pixel,
-                           cdelt2 * self.spatial_units[1] / u.pixel)
-
-    @property
-    def spatial_units(self):
-        """
-        Image coordinate units along the x and y axes (i.e. cunit1, cunit2).
-        """
-        units = self.meta.get('cunit1', None), self.meta.get('cunit2', None)
-        units = [None if unit is None else u.Unit(unit.lower()) for unit in units]
-        return SpatialPair(units[0], units[1])
-
-    @property
-    def rotation_matrix(self):
-        r"""
-        Matrix describing the transformation needed to align the reference
-        pixel with the coordinate axes.
-
-        The order or precedence of FITS keywords which this is taken from is:
-        - PC\*_\*
-        - CD\*_\*
-        - CROTA\*
-
-        Notes
-        -----
-        In many cases this is a simple rotation matrix, hence the property name.
-        It general it does not have to be a pure rotation matrix, and can encode
-        other transformations e.g., skews for non-orthgonal coordinate systems.
-        """
-        if any(key in self.meta for key in ['PC1_1', 'PC1_2', 'PC2_1', 'PC2_2']):
-            return np.array(
-                [
-                    [self.meta.get('PC1_1', 1), self.meta.get('PC1_2', 0)],
-                    [self.meta.get('PC2_1', 0), self.meta.get('PC2_2', 1)]
-                ]
-            )
-        elif any(key in self.meta for key in ['CD1_1', 'CD1_2', 'CD2_1', 'CD2_2']):
-            cd = np.array(
-                [
-                    [self.meta.get('CD1_1', 0), self.meta.get('CD1_2', 0)],
-                    [self.meta.get('CD2_1', 0), self.meta.get('CD2_2', 0)]
-                ]
-            )
-
-            cdelt = u.Quantity(self.scale).value
-
-            # Divide each row by each CDELT
-            return cd / np.expand_dims(cdelt, axis=1)
-        else:
-            return self._rotation_matrix_from_crota()
-
-    @staticmethod
-    def _pc_matrix(lam, angle):
-        """
-        Returns PC matrix from the scale ration (lam) and rotation
-        angle in radians (angle).
-        """
-        return np.array([[np.cos(angle), -1 * lam * np.sin(angle)],
-                         [1/lam * np.sin(angle), np.cos(angle)]])
-
-    def _rotation_matrix_from_crota(self, crota_key='CROTA2'):
-        """
-        This method converts the deprecated CROTA FITS kwargs to the new
-        PC rotation matrix.
-
-        This method can be overridden if an instruments header does not use this
-        conversion.
-
-        Parameters
-        ----------
-        crota_key : str, optional
-            The key to use for CROTA2. Defaults to 'CROTA2'.
-
-        Notes
-        -----
-        If the specified key isn't present in the metadata, a default rotation
-        of 0deg is returned.
-        """
-        lam = self.scale[1] / self.scale[0]
-        p = np.deg2rad(self.meta.get(crota_key, 0))
-        return self._pc_matrix(lam, p)
-
-    @property
-    def _pv_values(self):
-        """
-        Return any PV values in the metadata.
-        """
-        pattern = re.compile('pv[0-9]_[0-9]a', re.IGNORECASE)
-        pv_keys = [k for k in self.meta.keys() if pattern.match(k)]
-
-        pv_values = []
-        for k in pv_keys:
-            i, m = int(k[2]), int(k[4])
-            pv_values.append((i, m, self.meta[k]))
-        return pv_values
-
-    @property
-    def fits_header(self):
-        """
-        A `~astropy.io.fits.Header` representation of the ``meta`` attribute.
-        """
-        return sunpy.io._fits.header_to_fits(self.meta)
+        # Validate the WCS here.
+        w2.wcs.set()
+        return w2
 
 # #### Miscellaneous #### #
     def _get_cmap_name(self):
@@ -1404,43 +728,6 @@ class GenericMap(NDData):
         cmap_string = (self.observatory + self.detector +
                        str(int(self.wavelength.to('angstrom').value)))
         return cmap_string.lower()
-
-    def _validate_meta(self):
-        """
-        Validates some meta-information associated with a Map.
-
-        This method includes very basic validation checks which apply to
-        all of the kinds of files that sunpy can read. Datasource-specific
-        validation should be handled in the relevant file in the
-        sunpy.map.sources package.
-        """
-        msg = ('Image coordinate units for axis {} not present in metadata.')
-        err_message = []
-        for i in [0, 1]:
-            if self.spatial_units[i] is None:
-                err_message.append(msg.format(i+1, i+1))
-
-        if err_message:
-            err_message.append(
-                f'See {_META_FIX_URL} for instructions on how to add missing metadata.')
-            raise MapMetaValidationError('\n'.join(err_message))
-
-        for meta_property in ('waveunit', ):
-            if (self.meta.get(meta_property) and
-                u.Unit(self.meta.get(meta_property),
-                       parse_strict='silent').physical_type == 'unknown'):
-                warn_metadata(f"Unknown value for {meta_property.upper()}.")
-
-        if (self.coordinate_system[0].startswith(('SOLX', 'SOLY')) or
-                self.coordinate_system[1].startswith(('SOLX', 'SOLY'))):
-            warn_user("sunpy Map does not support three dimensional data "
-                      "and therefore cannot represent heliocentric coordinates. Proceed at your own risk.")
-
-        if not all(su.is_equivalent(u.arcsec) for su in self.spatial_units):
-            units = [su.to_string() for su in self.spatial_units]
-            raise MapMetaValidationError(
-                'Map only supports spherical coordinate systems with angular units '
-                f'(ie. equivalent to arcsec), but this map has units {units}')
 
 # #### Data conversion routines #### #
     def world_to_pixel(self, coordinate):
