+++ sunpy/map/sources/rhessi.py
@@ -1,6 +1,7 @@
 import astropy.units as u
 
-from sunpy.map.mapbase import GenericMap, SpatialPair
+from sunpy.map.mapbase import GenericMap
+from sunpy.map.mixins.mapmeta import SpatialPair
 
 __all__ = ['RHESSIMap']
